+++ bar.py
@@ -3,15 +3,18 @@ import altair as alt
 import pandas as pd
 from add_data import db_execute_fetch
 
+
 def app():
     st.title('Bar Chart')
     st.write('You can see Ranking here')
     stBarChart()
 
+
 def loadData():
     query = "select * from TweetInformation"
     df = db_execute_fetch(query, dbName="tweets", rdf=True)
     return df
+
 
 def barChart(data, title, X, Y):
     title = title.title()
@@ -19,6 +22,7 @@ def barChart(data, title, X, Y):
     msgChart = (alt.Chart(data).mark_bar().encode(alt.X(f"{X}:N", sort=alt.EncodingSortField(field=f"{Y}", op="values",
                 order='ascending')), y=f"{Y}:Q"))
     st.altair_chart(msgChart, use_container_width=True)
+
 
 def stBarChart():
     df = loadData()
