+++ lang.py
@@ -1,14 +1,15 @@
 import streamlit as st
 import numpy as np
 import plotly.express as px
-
-def app():
-    st.title('APP2')
-    st.write('Welcome to app2')
 import streamlit as st
 import altair as alt
 import pandas as pd
 from add_data import db_execute_fetch
+
+
+def app():
+    st.title('APP2')
+    st.write('Welcome to app2')
 
 def app():
     st.title('Languages')
@@ -16,10 +17,12 @@ def app():
     langPie()
     st.markdown("<p style='padding:10px; background-color:#000000;color:#00ECB9;font-size:16px;border-radius:10px;'>Section Break</p>", unsafe_allow_html=True)
 
+
 def loadData():
     query = "select * from TweetInformation"
     df = db_execute_fetch(query, dbName="tweets", rdf=True)
     return df
+
 
 def langPie():
     df = loadData()
